ok spike loop is mostly working except for 650000 spikes

+++ port/src/gpu/cu_synfire.cuh
@@ -22,7 +22,7 @@ class CUSynfire {
 
     static CUSynfire CreateCUSynfire( int nsize );
 
-    static CUSynfire CreateCUSynfire( int nsize, double dt, int num_trials, int trial_time );
+    static CUSynfire CreateCUSynfire( int nsize, double dt, int num_trials, double trial_time );
 
     //~ CTOR
     CUSynfire( SynfireParameters );
